COMBINE: Config.tsx
merged Audio.tsx and Record.tsx together into Config.tsx to make moving data easier
