+++ TCap/.gitignore
@@ -23,3 +23,6 @@ dist-ssr
 *.sln
 *.sw?
 
+#ffmpeg target
+/**/target
+
